% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trainingData.R
\name{build_trainingData}
\alias{build_trainingData}
\title{Construct Training Data}
\usage{
build_trainingData(data, refcpx)
}
\arguments{
\item{data}{A co-elution data matrix with proteins in rows and
fractions in columns.}

\item{refcpx}{A list of protein complexes to create class labels for
training set.}
}
\value{
A list containing training data, along with class labels.
}
\description{
This function creates a training set with class labels,
where "1" corresponds to positive interactions (i.e., interactions
belongs to  the same complex), and "0" corresponds to negative
interactions (i.e., random interactions). To avoid prediction biases,
this function creates a class labels with the same number of pairs. The
training set then can be used as input for model tuning and model training.
}
\details{
build_trainingData
}
\examples{
# Load the list of ground-truth complexes
data("refcpx")
# Load the co-elution data
data("HelaCE")
# Buid the training set
t_data <- build_trainingData(HelaCE, refcpx)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
