% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_tunning.R
\name{Clust_tuning}
\alias{Clust_tuning}
\title{ClusterONE Hyperparameters Tuning}
\usage{
Clust_tuning(
  refcpx,
  csize = 3,
  d = c(0.3, 0.5),
  p = c(2),
  mx_overlap = c(0.5, 0.6),
  tpath = file.path(system.file("extdata", package = "deepCE"))
)
}
\arguments{
\item{refcpx}{A list containing reference complexes
(i.e., corum complexes).}

\item{csize}{Numerical value, the minimum size of the predicted complexes.
Defualts to 3.}

\item{d}{A vector of number, density of predicted complexes.}

\item{p}{A vector of integer, penalty value for the inclusion of each node}

\item{mx_overlap}{A vector of number, specifies the maximum allowed
overlap between two clusters.}

\item{tpath}{A character string indicating the path to the project
directory that contains the interaction data. Interaction data must be
stored as .txt file and containing id1-id2-weight triplets.}
}
\value{
A data.frame containing clustering performance across different
combination of parameters.
}
\description{
This function optimize the choice of ClusterONE algorithm
parameters such as density, node penalty, and overlap score by comparing
clustering-derived partitions for each combination of parameters to known
labels (i.e., CORUM complexes) and assess the similarity between them
using quality measures including overlap score, sensitivity (Sn),
clustering-wise positive predictive value (PPV), geometric accuracy (Acc),
and maximum matching raio (MMR). It is recommended to first reduce
redundancy in the known reference complexes
via \code{\link{RemoveCpxRedundance}}, then performs parameter tunning.
}
\details{
Clust_tuning
}
\examples{
Clust_tuning_Result <-
Clust_tuning(refcpx,
csize = 3,
d = c(0.3, 0.5),
p = c(2),
mx_overlap = c(0.5,0.6),
tpath = file.path(system.file("extdata", package = "deepCE")))
}
\references{
Nepusz, T., Yu, H., and Paccanaro, A. (2012a).
Detecting overlapping protein complexes in protein-protein interaction
networks. Nat. Methods 9, 471.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
