% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPPI.R
\name{diffPPI}
\alias{diffPPI}
\title{Calculation of Differential Protein-Protein Interactions}
\usage{
diffPPI(
  m1,
  m2,
  term,
  p.adjust.methods = "local",
  pvalue = 0.05,
  r1_r2 = 0.5,
  minSize = 5,
  maxSize = 500,
  tpath = tempdir()
)
}
\arguments{
\item{m1}{a numeric matrix with proteins in rows and
fractions in columns}

\item{m2}{a numeric matrix with proteins in rows and
fractions in columns}

\item{term}{A list containing known complexes/Go annotation/pathways
provided by the user.}

\item{p.adjust.methods}{c("local", holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none"). Defaults to local.}

\item{pvalue}{An integer specifying the p-value cutoff for differential
interactions. Defaults to 0.05.}

\item{r1_r2}{A number specifying the cutoff for the correlation
difference between two conditions. All interactions above and below this
threshold are excluded. Defaults to |0.5|.}

\item{minSize}{Minimal size of a gene set to test.
All pathways below the threshold are excluded. Defaults to 5.
See \code{\link[fgsea]{fgsea}} for more details}

\item{maxSize}{Maximal size of a gene set to test. All pathways above
the threshold are excluded.
See \code{\link[fgsea]{fgsea}} for more details. Defaults to 500.}

\item{tpath}{A character string indicating the path to the project
directory. If the directory is
missing, it will be stored in the Temp directory.}
}
\value{
Return following data sets:
 \itemize{
 \item{differential_PPI_unfilt} - Unfiltered differential interactions.
 \item{differential_PPI_filtered} - High-confidence differential
 interactions defined by user-defined threshold.
 \item{fgseaRes} - fgsea enrichment terms for high-confidence differential
 PPIs
 \item{vec_rank} - Rank file
 \item{term_list} - term_list
 }.
}
\description{
This function first computes the differential correlations
of two correlation metrics via ‘DiffCorr’ R package. Identified DIPs can
then be mapped to complexes/Go annotation/pathways provided by the users
via gene set enrichment analysis (GSEA),
using the ‘fgsea’ R package to determine the biological meaning of DIPs.
}
\details{
diffPPI
}
\examples{
# profile 1
data("m1")
# profile 2
data("m2")
# biological term (here is known complexes)
data("refcpx")
diff_output <-
diffPPI(m1,
m2,
refcpx,
minSize = 2,
maxSize = 10000,
tpath = tempdir())
}
\references{
Fukushima, A. Gene (2013) 518, 209-214
}
\author{
Matineh Rahmatbakhsh
}
