% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ConsecutivePep.R
\name{filter_ConsecutivePep}
\alias{filter_ConsecutivePep}
\title{Filter Peptides by Consecutive Peptide Detection}
\usage{
filter_ConsecutivePep(data, min_stretch_length = 2)
}
\arguments{
\item{data}{A data matrix with rows including proteins and fractions
along the columns.}

\item{min_stretch_length}{Numeric integer, the minimal length a stretch of
continuous identifications has to have in order not to be removed.
Defaults to 2.}
}
\value{
Filtered matrix
}
\description{
This function removes peptides/proteins that have never been
detected in more than N consecutive fractions (here N = 2).
}
\details{
filter_ConsecutivePep
}
\examples{
# Load the data
data("HelaCE")
# Remover proteins that detected in more than 10 consecutive fractions
filt_dat <- filter_ConsecutivePep(HelaCE,10)
}
\references{
Bludau,I. et al. (2020) Complex-centric proteome profiling by SEC-SWATH-MS
for the parallel detection of hundreds of protein complexes.
Nat. Protoc., 15, 2341-2386.
}
