% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_refInt.R
\name{generate_refInt}
\alias{generate_refInt}
\title{Generate Class Labels for Data Input Based on Gold Reference Set}
\usage{
generate_refInt(data, refcpx)
}
\arguments{
\item{data}{A data frame with interacting proteins in the first two
columns.}

\item{refcpx}{A list containing gold standard protein complexes.}
}
\value{
A Data frame containing class labels for protein pairs in the
data input. Protein pairs involve in same protein complexes are assigned
to 1, otherwise 0.
}
\description{
This function creates class labels for protein pairs in the
same order in the data input based on gold reference set.
}
\details{
generate_refInt
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
