% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPI.R
\name{getPPI}
\alias{getPPI}
\title{Generating Protein-Protein Interaction (PPI) Vector}
\usage{
getPPI(data, similarity_calculation = FALSE, metric = "pearson")
}
\arguments{
\item{data}{A Co-elution data matrix with proteins in rows and
fractions in columns.}

\item{similarity_calculation}{If TRUE, this function first computes the
similarity of protein pairs based on co-elution pattern via one of the
user-defined metrics. It then removes protein pairs below 0.5
similarity score}

\item{metric}{The measure of association including:
c(pearson", "spearman", "kendall", "bicor", "cosine", "jaccard",
"euclidean", "manhattan", "weighted_rank", "mutual_information",
"zero_count_bayesian", "weighted_cross_correlation", or
"pearson_plusNoise"); Defaults to pearson.
See \code{\link{similarity_score}}.}
}
\value{
A concatenated matrix.
}
\description{
This function concatenates co-elution profiles for each
possible protein pairs to form a master matrix for input into the learner.
}
\details{
getPPI
}
\examples{
data("HelaCE")
conc_dat <- getPPI(HelaCE, similarity_calculation = TRUE,
metric = "pearson")
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
