% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_MissingData.R
\name{impute_MissingData}
\alias{impute_MissingData}
\title{Detect and Impute Missing Values}
\usage{
impute_MissingData(
  data,
  bound_left = 2,
  bound_right = 2,
  plotImputationSummary = FALSE,
  filename = "plots.pdf"
)
}
\arguments{
\item{data}{A data matrix with rows
including proteins and fractions along the columns.}

\item{bound_left}{Numeric integer, the minimum number of non-zero
values to the left of a missing value to be replaced with \code{NA};
defaults to 2.}

\item{bound_right}{Numeric integer, the minimum number of non-zero
values to the right of a missing value to be replaced with \code{NA};
defaults to 2.}

\item{plotImputationSummary}{Logical value, indicating whether to  plot
the imputation summary; defaults to FALSE.}

\item{filename}{Character string, indicating the output file name as an
pdf object.}
}
\value{
Imputed matrix.
}
\description{
The algorithm first identifies 0 values on the data set as
missing values if they fulfill the rule:
\code{(1)*bound_left - 0 - (1)*bound_right} i.e. a zero value has to have
at least a specified number of non-zero neighbors to be classified
as \code{NA}. NA values are then linearly extrapolated from the
neighboring  values via interpolation. Any imputed value below 0
is set to 0.
}
\details{
impute_MissingData
}
\examples{
# Load the data
data("HelaCE")
# Impute the missing values
Imputed_data <-
impute_MissingData(HelaCE,bound_left = 2,bound_right = 2)
}
\references{
Bludau,I. et al. (2020) Complex-centric proteome profiling by SEC-SWATH-MS
for the parallel detection of hundreds of protein complexes.
Nat. Protoc., 15, 2341-2386.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
