% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1.R
\docType{data}
\name{m1}
\alias{m1}
\title{Co-elution profiles from DNLCs}
\format{
A matrix with 877 rows and 170 columns,
with proteins in rows and biochemical fractions in columns
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6831851/}
}
\usage{
data(m1)
}
\description{
Subset of co-elution profiles derived from various
biochemical techniques of  retinoic acid (RA)-induced differentiated
neuronal-like cells (DNLCs).
}
