% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\docType{data}
\name{m2}
\alias{m2}
\title{Co-elution profiles from ECSCs}
\format{
A matrix with 475 rows and 179 columns,
with proteins in rows and biochemical fractions in columns
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6831851/}
}
\usage{
data(m2)
}
\description{
Subset of co-elution profiles derived from various
biochemical techniques of NTera2 embryonal carcinoma stem cells (ECSCs).
}
