% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDCNN_tunning.R
\name{oneDCNN_tunning}
\alias{oneDCNN_tunning}
\title{Tune Hyperparamters for 1D-CNN}
\usage{
oneDCNN_tunning(
  train_d,
  train_l,
  nlayers = 2,
  filters_1 = c(32, 64),
  filters_2 = c(32, 64),
  kernel_size = c(1, 2),
  strides = 2,
  pool_size = 2,
  powerto = 6,
  drate = 0.1,
  optimizer = "rmsprop",
  b_size = 64,
  metrics = "accuracy",
  epochs = 10,
  k = 0.1
)
}
\arguments{
\item{train_d}{A matrix of training data containing numerical features,
generated from \code{\link{build_trainingData}}.}

\item{train_l}{A vector of binary categorical label (0-1),
generated from  \code{\link{build_trainingData}}.}

\item{nlayers}{A vector of integers, describing
number of hidden layers; defaults to 3.}

\item{filters_1}{A vector of integers, describing
the dimensionality of the output space
(i.e. the number of output filters in the first convolution).}

\item{filters_2}{A vector of integers, describing the
dimensionality of the output space
(i.e. the number of output filters in the subsequent convolution).}

\item{kernel_size}{A vector of integers, specifying the
height and width of the 2D convolution window. Can be a single integer to
specify the same value for all spatial dimensions.}

\item{strides}{A vector of integers, specifying the
strides of the convolution along the height and width. Can be a single
integer to specify the same value for all spatial dimensions.}

\item{pool_size}{A vector of integers, describing the pool_size.}

\item{powerto}{A vector of integers, describing
the number of neurons in the last layer as defined by two to
the power of this value.}

\item{drate}{A vector of numbers, describing the dropout rates.}

\item{optimizer}{A vector of strings, specifying the name of optimizer.
For most models, this defaults to "rmsprop".}

\item{b_size}{A vector of integers, describing,
number of samples per gradient update.}

\item{metrics}{List of metrics to be used by the model during k-fold
cross-validation. Defaults to 'accuracy'.}

\item{epochs}{Number of epochs to train the model.}

\item{k}{Float between 0 and 1. Fraction of the training data to be used
as validation data.}
}
\value{
A data.frame containing model performance across different
combination of parameters.
}
\description{
This function tune different combinations of hyperparameters
for one-dimensional convolutional neural network (1D-CNN) 1D-CNN model.
}
\details{
oneDCNN_tunning
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
