% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveSinglePeak.R
\name{RemoveSinglePeak}
\alias{RemoveSinglePeak}
\title{Remove Single Peak}
\usage{
RemoveSinglePeak(data, plot_RemovalSinglePeaks = FALSE, filename = "plots.pdf")
}
\arguments{
\item{data}{A data matrix with rows including proteins and fractions
along the columns.}

\item{plot_RemovalSinglePeaks}{Logical value, indicating whether to  plot
the distribution after peak removal; Defaults to FALSE.}

\item{filename}{Character string, indicating the output file name as an
pdf object.}
}
\value{
Filtered matrix.
}
\description{
This function replaces single values surrounded by zeroes
with zero, eliminating data-noise “peaks” that only span one fraction.
}
\details{
RemoveSinglePeak
}
\examples{
data("HelaCE")
filt_data <- RemoveSinglePeak(HelaCE)
}
\references{
Schlossarek,D. et al. (2021) PROMISed: A novel web-based tool to
facilitate analysis and visualization of the molecular interaction
networks from co-fractionation mass spectrometry (CF-MS) experiments.
Comput. Struct. Biotechnol.
}
