% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_m.R
\name{scaling_m}
\alias{scaling_m}
\title{Row-wise Normalization}
\usage{
scaling_m(data)
}
\arguments{
\item{data}{A data matrix with rows
including proteins and fractions along the columns.}
}
\value{
Scaled data matrix.
}
\description{
This function performs row-wise normalization by dividing
the ion intensity or spectral count of the particular protein by the
max of ion intensities or spectra counts across fractions.
}
\details{
scaling_m
}
\examples{
data("HelaCE")
n_data <- scaling_m(HelaCE)
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
