% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_score.R
\name{similarity_score}
\alias{similarity_score}
\title{Calculate Pairwise Protein Profile Similarity using Different
  Metrics}
\usage{
similarity_score(mat, metric = "pearson", rept = 10)
}
\arguments{
\item{mat}{A Co-elution data matrix with proteins in rows and fractions in
columns.}

\item{metric}{The measure of association including: c(pearson",
"spearman", "kendall", "bicor", "cosine", "jaccard", "euclidean",
"manhattan", "weighted_rank", "mutual_information",
"zero_count_bayesian", "weighted_cross_correlation", or
"pearson_plusNoise"); Defaults to pearson.}

\item{rept}{Poisson iterations for pearson_plusNoise metric, defaults to
10.}
}
\value{
a matrix of dimensions (# of proteins) x (# of proteins), scoring
  every possible protein pair, in which higher values reflect more similar
  pairs.
}
\description{
This function computes pairwise protein similarity in a given
 CF-MS experiment using up to 13 indices.
}
\details{
similarity_score
}
\examples{
M1<-matrix(rnorm(36),nrow=6)
M1 <- abs(M1)
rownames(M1) <- c("A","B","C","D","E","F")
scored_Data <- similarity_score(M1, metric = "euclidean")
}
\references{
Skinnider,M.A. and Foster,L.J. (2021) Meta-analysis defines
  principles for the design and analysis of co-fractionation mass
  spectrometry experiments. Nat. Methods.
}
\author{
Matineh Rahmatbakhsh, \email{matinerb.94@gmail.com}
}
